ftom riak_dt_pncounter to antidote_crdt_counter
We got rid of a proto file in the depencencies that was causing trouble
with antidote_pb_codec, but that means losing a few datatypes.

Now we can only use these types that can be matched in antidote_pb_codec

But we do have to have a version of riak_pb in _checkouts, because
otherwise it won't compile anymore.